+++ src/components/BsField/index.js
@@ -1,5 +1,3 @@
-import BsCheckbox from "./BsCheckbox";
-import BsCheckboxGroup from "./BsCheckboxGroup";
 import BsCombobox from "./BsCombobox";
 import BsTextField from "./BsTextField";
 import BsTextArea from "./BsTextArea";
@@ -7,13 +5,9 @@ import BsNumericField from "./BsNumericField";
 import BsDateTimeField from "./BsDateTimeField";
 import BsChipField from "./BsChipField";
 import BsSearchField from "./BsSearchField";
-import BsRadio from "./BsRadio";
-import BsRadioGroup from "./BsRadioGroup";
 import BsSwitch from "./BsSwitch";
 
 export default Vue => {
-    Vue.component(BsCheckbox.name, BsCheckbox);
-    Vue.component(BsCheckboxGroup.name, BsCheckboxGroup);
     Vue.component(BsCombobox.name, BsCombobox);
     Vue.component(BsTextField.name, BsTextField);
     Vue.component(BsTextArea.name, BsTextArea);
@@ -21,7 +15,5 @@ export default Vue => {
     Vue.component(BsDateTimeField.name, BsDateTimeField);
     Vue.component(BsChipField.name, BsChipField);
     Vue.component(BsSearchField.name, BsSearchField);
-    Vue.component(BsRadio.name, BsRadio);
-    Vue.component(BsRadioGroup.name, BsRadioGroup);
     Vue.component(BsSwitch.name, BsSwitch);
 };