+++ src/components/Checkbox/BsCheckbox.ts
@@ -1,8 +1,9 @@
 import {ComponentOptionsMixin, computed, ComputedOptions, defineComponent, EmitsOptions, nextTick, ref} from "vue";
-import {useCheckboxClasses, useRenderCheckbox} from "./mixins/checkboxApi";
+import {useCheckboxClasses, useCheckSelected, useCreateCheckboxElement} from "./mixins/checkboxApi";
 import {checkboxProps} from "./mixins/checkboxProps";
 import {TBsCheckbox, TCheckboxOptionProps} from "./types";
 import {TRecord} from "../../types";
+import {useRenderRadioOrCheckbox} from "../Radio/mixins/radioApi";
 
 export default defineComponent<TBsCheckbox, TRecord, TRecord, ComputedOptions, ComponentOptionsMixin, EmitsOptions>({
     name: "BsCheckbox",
@@ -20,14 +21,12 @@ export default defineComponent<TBsCheckbox, TRecord, TRecord, ComputedOptions, C
     setup(props, {emit, slots}) {
         const cmpProps = props as Readonly<TCheckboxOptionProps>;
         const rippleActive = ref<boolean>(false);
-        const checkboxClasses = computed(
-            () => useCheckboxClasses(cmpProps)
-        )
+        const checkboxClasses = computed(() => useCheckboxClasses(cmpProps));
         const toggleCheckHandler = (): void => {
             if (!cmpProps.disabled && !cmpProps.readonly) {
-                const checked = cmpProps.value === cmpProps.modelValue;
+                const checked = useCheckSelected(cmpProps);
                 rippleActive.value = true;
-                emit("update:modelValue", checked ? null : cmpProps.value)
+                emit("update:modelValue", (checked ? null : cmpProps.value))
                 nextTick().then(() => {
                     emit("checked", !checked);
                 });
@@ -35,9 +34,9 @@ export default defineComponent<TBsCheckbox, TRecord, TRecord, ComputedOptions, C
         }
 
         return () =>
-            useRenderCheckbox(
-                slots, cmpProps, checkboxClasses, rippleActive,
-                toggleCheckHandler,
+            useRenderRadioOrCheckbox(
+                slots, cmpProps, checkboxClasses, rippleActive, "checkbox",
+                useCreateCheckboxElement(cmpProps), toggleCheckHandler,
             );
     }
 });